[PATCH] procfs: fix printk arg type warning
On sparc32 build, there is a printk format arg-type warning:
fs/proc/proc_misc.c:195: warning: long unsigned int format, unsigned int arg (arg 23)

I tried to fix it with a change to asm-sparc/vaddrs.h:
-#define VMALLOC_START          0xfe600000
+#define VMALLOC_START          0xfe600000UL
-#define VMALLOC_END            0xffc00000
+#define VMALLOC_END            0xffc00000UL

but that won't fly because the #defines are used in asm code and asm doesn't
like the UL suffixes (reported by Bill Irwin).

+++ fs/proc/proc_misc.c
@@ -189,7 +189,7 @@ static int meminfo_read_proc(char *page, char **start, off_t off,
 		K(allowed),
 		K(committed),
 		K(ps.nr_page_table_pages),
-		VMALLOC_TOTAL >> 10,
+		(unsigned long)VMALLOC_TOTAL >> 10,
 		vmi.used >> 10,
 		vmi.largest_chunk >> 10
 		);
